requirements.txt
fastapi
sqlmodel
psycopg2
python-dotenv
uvicorn
